+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.0.11] - 2026-08-11
 
 MySQL quest progress display fix.